Create a fake 'white' to prevent Prince making white transparent
I think that to Prince 'white' is the paper colour, so it's
transparent. But on a grey background, when Prince applies PDF X1a
and removes transparency, that white/transparent disappears.
To prevent this, I've created a colour called $white that's
actually just 1% black, and used it for icons on grey. The 1% black
should not be visible when printed.